Main updated & vampires fixed
