--- src/org/ucm/tp1/BuffyVampireSlayer.java
@@ -1,5 +1,0 @@
-package org.ucm.tp1;
-
-public class BuffyVampireSlayer {
-
-}
